+++ packages/i18nify-js/src/modules/geo/constants.ts
@@ -1,1 +1,3 @@
+export const FLAG_4X3_BASE_PATH =
+  'https://unpkg.com/@razorpay/i18nify-js/lib/assets/flags';
 export const FLAG_BASE_PATH = 'https://flagcdn.com';