+++ packages/i18nify-js/src/modules/geo/getFlagsForAllCountries.ts
@@ -1,7 +1,7 @@
 import { withErrorBoundary } from '../../common/errorBoundary';
 import { LIST_OF_ALL_COUNTRIES } from './data/listOfAllCountries';
-import { FLAG_BASE_PATH } from './constants';
-import { CountryCodeType } from '../types';
+import { FLAG_4X3_BASE_PATH, FLAG_BASE_PATH } from './constants';
+import { CountryCodeType, GetFlagReturnType } from '../types';
 
 /**
  * Retrieves a mapping of country codes to their corresponding flag image URLs.
@@ -13,17 +13,20 @@ import { CountryCodeType } from '../types';
  * @returns An object mapping each country code from the list to its flag image URL.
  */
 const getFlagsForAllCountries = (): {
-  [countryCode in CountryCodeType]: string;
+  [countryCode in CountryCodeType]: GetFlagReturnType;
 } => {
   // Initialize an empty object to hold the country code to flag URL mapping
-  const flagsForAllCountriesMap: { [countryCode in CountryCodeType]: string } =
-    {} as { [countryCode in CountryCodeType]: string };
+  const flagsForAllCountriesMap = {} as {
+    [countryCode in CountryCodeType]: GetFlagReturnType;
+  };
 
   // Loop through each country code in the list
   LIST_OF_ALL_COUNTRIES.map((countryCode: CountryCodeType) => {
     // Construct the flag image URL and assign it to the corresponding country code in the map
-    flagsForAllCountriesMap[countryCode] =
-      `${FLAG_BASE_PATH}/${countryCode.toLowerCase()}.svg`;
+    flagsForAllCountriesMap[countryCode] = {
+      original: `${FLAG_BASE_PATH}/${countryCode}.svg`,
+      '4X3': `${FLAG_4X3_BASE_PATH}/${countryCode}.svg`,
+    };
   });
 
   // Return the populated map of country codes to flag image URLs